+++ velocity/src/main/java/com/mrivanplays/conversations/velocity/VelocityTimeoutScheduler.java
@@ -18,7 +18,7 @@ class VelocityTimeoutScheduler implements TimeoutScheduler {
   }
 
   @Override
-  public TimeoutTask schedule(Runnable task, int time, TimeUnit timeUnit) {
+  public TimeoutTask schedule(Runnable task, long time, TimeUnit timeUnit) {
     return new VelocityTimeoutTask(
         proxy.getScheduler().buildTask(plugin, task).delay(time, timeUnit).schedule());
   }
